+++ pysatNASA/instruments/methods/cdaweb.py
@@ -16,6 +16,7 @@ import warnings
 import xarray as xr
 
 from bs4 import BeautifulSoup
+from cdasws import CdasWs
 
 import pysat
 from pysat.instruments.methods import general
@@ -29,6 +30,35 @@ try:
     auto_CDF = pysatCDF.CDF
 except ImportError:
     auto_CDF = libCDF
+
+
+def try_inst_dict(inst_id, tag, supported_tags):
+    """Check that the inst_id and tag combination is valid.
+
+    Parameters
+    ----------
+    tag : str
+        Data product tag (default='')
+    inst_id : str
+        Instrument ID (default='')
+    supported_tags : dict
+        dict of dicts. Keys are supported tag names for download. Value is
+        a dict with 'remote_dir', 'fname'. Inteded to be
+        pre-set with functools.partial then assigned to new instrument code.
+        (default=None)
+
+    Returns
+    -------
+    inst_dict : dict or str
+        dictionary containing file location in spdf archive, or dataset ID for
+        cdasws
+    """
+    try:
+        inst_dict = supported_tags[inst_id][tag]
+    except KeyError:
+        raise ValueError('inst_id / tag combo unknown.')
+
+    return inst_dict
 
 
 def load(fnames, tag='', inst_id='', file_cadence=dt.timedelta(days=1),
@@ -439,10 +469,7 @@ def download(date_array, tag='', inst_id='', supported_tags=None,
 
     """
 
-    try:
-        inst_dict = supported_tags[inst_id][tag]
-    except KeyError:
-        raise ValueError('inst_id / tag combo unknown.')
+    inst_dict = try_inst_dict(inst_id, tag, supported_tags)
 
     # Naming scheme for files on the CDAWeb server
     remote_dir = inst_dict['remote_dir']
@@ -494,6 +521,101 @@ def download(date_array, tag='', inst_id='', supported_tags=None,
         except requests.exceptions.RequestException as exception:
             logger.info(' '.join((str(exception), '- File not available for',
                                   date.strftime('%d %B %Y'))))
+    return
+
+
+def cdas_download(date_array, tag='', inst_id='', supported_tags=None,
+                  data_path=None):
+    """Download NASA CDAWeb CDF data using cdasws.
+
+    This routine is intended to be used by pysat instrument modules supporting
+    a particular NASA CDAWeb dataset.
+
+    Parameters
+    ----------
+    date_array : array-like
+        Array of datetimes to download data for. Provided by pysat.
+    tag : str
+        Data product tag (default='')
+    inst_id : str
+        Instrument ID (default='')
+    supported_tags : dict
+        dict of dicts. Keys are supported tag names for download. Value is
+        a dict with 'remote_dir', 'fname'. Inteded to be pre-set with
+        functools.partial then assigned to new instrument code.
+        (default=None)
+    data_path : str or NoneType
+        Path to data directory.  If None is specified, the value previously
+        set in Instrument.files.data_path is used.  (default=None)
+
+    Note
+    ----
+    Supported tags for this function use the cdaweb dataset naming convention.
+    You can find the data set names on CDAWeb or you can use cdasws.
+
+    Starting from scratch using cdasws
+    ::
+        from cdasws import CdasWs
+        cdas = CdasWs()
+
+        # Get list of available observatories/platforms.
+        cdas.get_observatories()
+
+        # Once your observatory is located, list the available instruments.
+        cdas.get_instruments(observatory=‘observatory_name’)
+
+        # Now list the available data sets for one instrument.
+        cdas.get_datasets(observatory=‘observatory_name’,
+                          instrument=‘instrument_name’)
+
+        # You can also list all of the datasets for an observatory.
+        cdas.get_datasets(observatory=‘observatory_name’)
+
+    Alternatively
+    ::
+        Visit https://cdaweb.gsfc.nasa.gov/
+        Select the observatory you want from the list and press submit.
+        The following page will have a list of the data sets.
+        The bolded names are in the format that cdasws uses.
+
+    Examples
+    --------
+    ::
+        # Download support added to cnofs_vefi.py using code below
+        download_tags = {'': {'dc_b': 'CNOFS_VEFI_BFIELD_1SEC'}}
+        download = functools.partial(cdw.cdas_download,
+                                     supported_tags=download_tags)
+
+    """
+
+    start = date_array[0]
+    stop = date_array[-1]
+    remote_files = cdas_list_remote_files(tag=tag, inst_id=inst_id,
+                                          start=start, stop=stop,
+                                          supported_tags=supported_tags,
+                                          series_out=False)
+
+    for file in remote_files:
+
+        fname = file.split('/')[-1]
+        saved_local_fname = os.path.join(data_path, fname)
+
+        # Perform download
+        logger.info(' '.join(('Attempting to download file: ',
+                              file)))
+        try:
+            with requests.get(file) as req:
+                if req.status_code != 404:
+                    with open(saved_local_fname, 'wb') as open_f:
+                        open_f.write(req.content)
+                    logger.info('Successfully downloaded {:}.'.format(
+                        saved_local_fname))
+                else:
+                    logger.info(' '.join(('File: "', file,
+                                          '" is not available')))
+        except requests.exceptions.RequestException as exception:
+            logger.info(' '.join((str(exception), '- File: "', file,
+                                  '" Is not available')))
     return
 
 
@@ -560,10 +682,7 @@ def list_remote_files(tag='', inst_id='', start=None, stop=None,
 
     """
 
-    try:
-        inst_dict = supported_tags[inst_id][tag]
-    except KeyError:
-        raise ValueError('inst_id / tag combo unknown.')
+    inst_dict = try_inst_dict(inst_id, tag, supported_tags)
 
     # Naming scheme for files on the CDAWeb server
     format_str = '/'.join((inst_dict['remote_dir'].strip('/'),
@@ -672,3 +791,82 @@ def list_remote_files(tag='', inst_id='', start=None, stop=None,
         stored_list = stored_list[mask]
 
     return stored_list
+
+
+def cdas_list_remote_files(tag='', inst_id='', start=None, stop=None,
+                           supported_tags=None, series_out=True):
+    """Return a list of every file for chosen remote data.
+
+    This routine is intended to be used by pysat instrument modules supporting
+    a particular NASA CDAWeb dataset.
+
+    Parameters
+    ----------
+    tag : str
+        Data product tag (default='')
+    inst_id : str
+        Instrument ID (default='')
+    start : dt.datetime or NoneType
+        Starting time for file list. A None value will start with the first
+        file found.
+        (default=None)
+    stop : dt.datetime or NoneType
+        Ending time for the file list.  A None value will stop with the last
+        file found.
+        (default=None)
+    supported_tags : dict
+        dict of dicts. Keys are supported tag names for download. Value is
+        a dict with 'remote_dir', 'fname'. Inteded to be
+        pre-set with functools.partial then assigned to new instrument code.
+        (default=None)
+    series_out : bool
+        boolean to determine output type. True for pandas series of file names,
+        and False for a list of the full web address.
+
+    Returns
+    -------
+    file_list : list
+        A list containing the verified available files
+
+    Note
+    ----
+    Supported tags for this function use the cdaweb dataset naming convention.
+    You can find the dataset names on cdaweb or you can use cdasws.
+
+    Examples
+    --------
+    ::
+        download_tags = {'': {'dc_b': 'CNOFS_VEFI_BFIELD_1SEC'}}
+        list_remote_files = functools.partial(cdw.cdas_list_remote_files,
+                                              supported_tags=download_tags)
+
+        download_tags = {'': {'': 'CNOFS_CINDI_IVM_500MS'}}
+        list_remote_files = functools.partial(cdw.cdas_list_remote_files,
+                                              supported_tags=download_tags)
+    """
+    cdas = CdasWs()
+    dataset = try_inst_dict(inst_id, tag, supported_tags)
+
+    if start is None and stop is None:
+        # Use the topmost directory without variables
+        start = cdas.get_inventory(identifier=dataset)[0].start
+        stop = cdas.get_inventory(identifier=dataset)[-1].end
+    elif stop is None:
+        stop = start + dt.timedelta(days=1)
+    elif start == stop:
+        stop = start + dt.timedelta(days=1)
+
+    if isinstance(start, pds._libs.tslibs.timestamps.Timestamp):
+        start = start.tz_localize('utc')
+        stop = stop.tz_localize('utc')
+
+    og_files = cdas.get_original_files(dataset=dataset, start=start, end=stop)
+
+    if series_out:
+        name_list = [os.path.basename(f['Name']) for f in og_files[1]]
+        t_stamp = [pds.Timestamp(f['StartTime'][:10]) for f in og_files[1]]
+        file_list = pds.Series(data=name_list, index=t_stamp)
+    else:
+        file_list = [f['Name'] for f in og_files[1]]
+
+    return file_list