+++ CHANGELOG.md
@@ -9,7 +9,11 @@ This project adheres to [Semantic Versioning](https://semver.org/).
   * ACE SIS
   * ACE SWEPAM
   * TIMED GUVI L1C intensity data and L2 EDR-Aurora data
+    * Type of sensor source handled by inst_id with options of
+      spectrograph, imaging
+    * Resolution of dataset handled by tag with low, high
   * DMSP SSUSI EDR-Aurora data
+* Added CDAWeb methods that can use cdasws to get the remote file list
 * Bug Fixes
   * Updated CDAWeb routines to allow for data stored by year/day-of-year
   * Added missing sub-module imports